Dedup positions
When generating a list of all the positions we care about, it's possible
that we can end up with duplicates (e.g. if two of the positions we want
have subsequently been merged at Wikidata), which will end up with
someone listed as holding the same position twice.

To work around this, dedup the generated master list.