Archive raw wtf_wikipedia output
Store the raw wtf_wikipedia output separately from the processed infobox
data.

The first can be directly committed: this is good to have in each repo,
and once it's there we can start to run other processing on it, but we
don't need to be actively tracking changes to it.

The second should trigger a PR, so that changes can be investigated.